pneumonia/requirements.txt
Flask
Flask-Cors
gunicorn
tensorflow
tensorflow_hub
numpy
pillow
